initial work on tmux terminator integration
some fixes, some refactoring

start handling terminal size

refactor some

add a conditional for tmux mode

refractor + add window layout parsing

refactor some more

fix

initial work on split window

add support for terminal close

close window on ui thread

add debug log on failed requests

add name to todos

add initial glue code for tmux based layouting

more layout prep boilerplate

disable refresh client for now

refactor request mechanism

add initial implementation of conversion between tmux layouts to terminator layouts

initial work on actual tmux based layouting

consume output on initial attach

minor output improvement

preserve (most) whitespace in initial output

small fix

bug fixes

+++ terminatorlib/terminal.py
@@ -108,6 +108,9 @@ class Terminal(Gtk.VBox):
     cnxids = None
     targets_for_new_group = None
 
+    control = None
+    pane_id = None
+
     def __init__(self):
         """Class initialiser"""
         GObject.GObject.__init__(self)
@@ -176,6 +179,8 @@ class Terminal(Gtk.VBox):
         self.reconfigure()
         self.vte.set_size(80, 24)
 
+        self.control = self.terminator.tmux_control
+
     def get_vte(self):
         """This simply returns the vte widget we are using"""
         return(self.vte)
@@ -229,14 +234,15 @@ class Terminal(Gtk.VBox):
         dbg('close: called')
         self.cnxids.remove_widget(self.vte)
         self.emit('close-term')
-        try:
-            dbg('close: killing %d' % self.pid)
-            os.kill(self.pid, signal.SIGHUP)
-        except Exception, ex:
-            # We really don't want to care if this failed. Deep OS voodoo is
-            # not what we should be doing.
-            dbg('os.kill failed: %s' % ex)
-            pass
+        if not util.TMUX:
+            try:
+                dbg('close: killing %d' % self.pid)
+                os.kill(self.pid, signal.SIGHUP)
+            except Exception, ex:
+                # We really don't want to care if this failed. Deep OS voodoo
+                # is not what we should be doing.
+                dbg('os.kill failed: %s' % ex)
+                pass
 
         if self.vte:
             self.terminalbox.remove(self.vte)
@@ -898,6 +904,9 @@ class Terminal(Gtk.VBox):
             if groupsend == groupsend_type['all']:
                 self.terminator.all_emit(self, 'key-press-event', event)
 
+        if util.TMUX:
+            self.control.send_keypress(event, pane_id=self.pane_id)
+
         return(False)
 
     def on_buttonpress(self, widget, event):
@@ -1261,8 +1270,11 @@ class Terminal(Gtk.VBox):
         self.on_vte_size_allocate(widget, allocation)
 
     def on_vte_size_allocate(self, widget, allocation):
-        self.titlebar.update_terminal_size(self.vte.get_column_count(),
-                self.vte.get_row_count())
+        column_count = self.vte.get_column_count()
+        row_count = self.vte.get_row_count()
+        self.titlebar.update_terminal_size(column_count, row_count)
+        # if util.TMUX:
+        #     self.control.refresh_client(column_count, row_count)
         if self.config['geometry_hinting']:
             window = self.get_toplevel()
             window.deferred_set_rough_geometry_hints()
@@ -1353,7 +1365,8 @@ class Terminal(Gtk.VBox):
         if cwd is not None:
             self.cwd = cwd
 
-    def spawn_child(self, widget=None, respawn=False, debugserver=False):
+    def spawn_child(self, widget=None, respawn=False, debugserver=False,
+                    orientation=None, active_pane_id=None):
         args = []
         shell = None
         command = None
@@ -1425,15 +1438,28 @@ class Terminal(Gtk.VBox):
             envv.append('TERMINATOR_DBUS_PATH=%s' % self.terminator.dbus_path)
 
         dbg('Forking shell: "%s" with args: %s' % (shell, args))
-        args.insert(0, shell)
-        result,  self.pid = self.vte.spawn_sync(Vte.PtyFlags.DEFAULT,
-                                       self.cwd,
-                                       args,
-                                       envv,
-                                       GLib.SpawnFlags.FILE_AND_ARGV_ZERO | GLib.SpawnFlags.DO_NOT_REAP_CHILD,
-                                       None,
-                                       None,
-                                       None)
+        if util.TMUX:
+            if self.terminator.initial_layout:
+                pass
+            else:
+                # command = '{} {}'.format(shell, ' '.join(args))
+                command = ' '.join(args)
+                self.pane_id = str(util.make_uuid())
+                self.control.run_command(command=command,
+                                         cwd=self.cwd,
+                                         marker=self.pane_id,
+                                         orientation=orientation,
+                                         pane_id=active_pane_id)
+        else:
+            args.insert(0, shell)
+            result,  self.pid = self.vte.spawn_sync(Vte.PtyFlags.DEFAULT,
+                                           self.cwd,
+                                           args,
+                                           envv,
+                                           GLib.SpawnFlags.FILE_AND_ARGV_ZERO | GLib.SpawnFlags.DO_NOT_REAP_CHILD,
+                                           None,
+                                           None,
+                                           None)
         self.command = shell
 
         self.titlebar.update()
@@ -1503,12 +1529,18 @@ class Terminal(Gtk.VBox):
 
     def paste_clipboard(self, primary=False):
         """Paste one of the two clipboards"""
-        for term in self.terminator.get_target_terms(self):
-            if primary:
-                term.vte.paste_primary()
-            else:
-                term.vte.paste_clipboard()
-        self.vte.grab_focus()
+        if util.TMUX:
+            def callback(_, content):
+                content = content.replace('\n',  '\r')
+                self.control.send_content(content, self.pane_id)
+            self.clipboard.request_text(callback)
+        else:
+            for term in self.terminator.get_target_terms(self):
+                if primary:
+                    term.vte.paste_primary()
+                else:
+                    term.vte.paste_clipboard()
+            self.vte.grab_focus()
 
     def feed(self, text):
         """Feed the supplied text to VTE"""
@@ -1644,6 +1676,11 @@ class Terminal(Gtk.VBox):
             self.directory = layout['directory']
         if layout.has_key('uuid') and layout['uuid'] != '':
             self.uuid = make_uuid(layout['uuid'])
+        if layout.has_key('tmux'):
+            tmux = layout['tmux']
+            self.pane_id = tmux['pane_id']
+            self.terminator.pane_id_to_terminal[self.pane_id] = self
+            self.control.initial_output(self.pane_id)
 
     def scroll_by_page(self, pages):
         """Scroll up or down in pages"""
